+++ src/main/java/io/khasang/moika/service/CompanyService.java
@@ -8,9 +8,9 @@ import java.util.List;
 public interface CompanyService {
     void addCompany(Company company);
     void addCompany(Company company, String name);
-    Company getCompanyById(long id);
+    Company getCompanyById(int id);
     void updateCompany(Company company);
-    void deleteCompany(long id);
+    void deleteCompany(int id);
     List<Company> getCompanyGazpromList();
     void addButterfly(Butterfly butterfly);
 }
